working on delete script

+++ cgi-bin/delete.sh
@@ -1,16 +1,32 @@
 #!/usr/bin/env bash
 
-rm $FILENAME 2> /dev/null > /dev/null 
-
-status=`echo $?`
-
-if [ $status -eq 0 ]
+if [ -z "$QUERY_STRING"  && -z "$1"]
 then
-	message="File deleted successfully"
-	status="200"
+	message="No file specified"
+	status="400"
 else
-	message="File could not be deleted"
-	status="500"
+	if [ -z $1 ]
+	then
+		FILENAME=$(echo "$QUERY_STRING" | sed -n 's/.*=\([^&]*\).*/\1/p' | sed 's/%20/ /g')
+	else
+		FILENAME="$1"
+	fi
+fi
+
+if [ ! -z $status ]
+then
+	rm $FILENAME 2> /dev/null > /dev/null 
+
+	status=`echo $?`
+
+	if [ $status -eq 0 ]
+	then
+		message="File deleted successfully"
+		status="200"
+	else
+		message="File could not be deleted"
+		status="500"
+	fi
 fi
 
 date=`date -u +"%a, %d %b %Y %H:%M:%S GMT"`
@@ -28,7 +44,9 @@ Content-Length: $content_length\r\n\
 Content-type: text/html; charset=utf-8\r\n\
 Date: $date\r\n\
 Last-Modified: $date\r\n\
+Connection: close\r\n\
 Server: CODAM_WEBSERV\r\n\r"
+# Server: "$HOST"\r\n\r"
 
 echo -e "$header"
 echo -e "$body"
